+++ docs/pages/release_notes.md
@@ -39,6 +39,7 @@ This is a {{ site.pmd.release_type }} release.
     *   [#1632](https://github.com/pmd/pmd/issues/1632): \[java] ConsecutiveLiteralAppends false positive over catch
 *   plsql
     *   [#1587](https://github.com/pmd/pmd/issues/1587): \[plsql] Parse Exception with EXISTS
+    *   [#1589](https://github.com/pmd/pmd/issues/1589): \[plsql] ParseException with subqueries in WHERE clause
 
 ### API Changes